tdf#99784 OSX run GUI stuff in the main thread
The extension manager starts dialogs from its own thread. But the
OSX backend currently doesn't defer these calls to the main thread.
This implements the deference using dispatch_*.

Since dispatch_* can't return any values, we're using a thread-id
based map to transfer the return values from the main thread to the
callers.

One major problem is the missing equivalent of Windows
WaitForMultipleObjects function, which we have to emulate by
try-locking the SolarMutex and waiting for an unlock message. Using
an unprotected boolean isn't race free, therefore we use a 0.1s
timout in the event lookup function. Our out-of-order SendMessage
processing works fine with the processing of queued dispatch_* code
blocks while waiting for new events.

Then there is the problem of the notifications, where we have to
disable the SolarMutex locking, because we generate these events
ourself, but must not release it. Just ignoring the notifications
currently results in GUI update failures.

The runinmain implementation uses macros, so it can rely on guard
objects for async calls.

Change-Id: Id8977991e3eda91da27c23d8021e028d4f4cefe5